+++ tests/test_listing.py
@@ -4,6 +4,7 @@ from codequick import listing, route, resolver
 from codequick.support import dispatcher
 from codequick.utils import unicode_type
 import xbmcgui
+import pickle
 import xbmc
 import sys
 
@@ -61,8 +62,8 @@ class Params(unittest.TestCase):
 
 class Art(Params):
     def setUp(self):
-        listitem = xbmcgui.ListItem()
-        self.base = listing.Art(listitem)
+        self.listitem = xbmcgui.ListItem()
+        self.base = listing.Art()
 
     def test_empty_setter(self):
         self.base["test"] = ""
@@ -83,7 +84,7 @@ class Art(Params):
         self.assertNotIn("thumb", self.base)
         self.assertNotIn("fanart", self.base)
         self.assertNotIn("icon", self.base)
-        self.base._close(False)
+        self.base._close(self.listitem, False)
         self.assertIn("thumb", self.base)
         self.assertIn("fanart", self.base)
         self.assertIn("icon", self.base)
@@ -92,7 +93,7 @@ class Art(Params):
         listing.fanart = "fanart"
         self.base["thumb"] = ""
         self.base["icon"] = ""
-        self.base._close(False)
+        self.base._close(self.listitem, False)
         self.assertNotIn("icon", self.base)
         self.assertNotIn("thumb", self.base)
         self.assertIn("fanart", self.base)
@@ -100,8 +101,8 @@ class Art(Params):
 
 class Info(Params):
     def setUp(self):
-        listitem = xbmcgui.ListItem()
-        self.base = listing.Info(listitem)
+        self.listitem = xbmcgui.ListItem()
+        self.base = listing.Info()
 
     def test_empty_setter(self):
         self.base["test"] = ""
@@ -166,17 +167,17 @@ class Info(Params):
 
     def test_close(self):
         self.base["plot"] = "plot"
-        self.base._close("video")
+        self.base._close(self.listitem, "video")
 
 
 class Property(Params):
     def setUp(self):
-        listitem = xbmcgui.ListItem()
-        self.base = listing.Property(listitem)
+        self.listitem = xbmcgui.ListItem()
+        self.base = listing.Property()
 
     def test_close(self):
         self.base["StartOffset"] = "256.4"
-        self.base._close()
+        self.base._close(self.listitem)
 
     def test_empty_setter(self):
         self.base["test"] = ""
@@ -185,8 +186,8 @@ class Property(Params):
 
 class Stream(Params):
     def setUp(self):
-        listitem = xbmcgui.ListItem()
-        self.base = listing.Stream(listitem)
+        self.listitem = xbmcgui.ListItem()
+        self.base = listing.Stream()
 
     def test_empty_setter(self):
         self.base["test"] = ""
@@ -257,18 +258,18 @@ class Stream(Params):
         self.base["video_codec"] = "h265"
         self.base["audio_language"] = "en"
         self.base["subtitle_language"] = "en"
-        self.base._close()
+        self.base._close(self.listitem)
 
     def test_close_invalid(self):
         self.base["subtitle_languages"] = "en"
         with self.assertRaises(KeyError):
-            self.base._close()
+            self.base._close(self.listitem)
 
 
 class Context(unittest.TestCase):
     def setUp(self):
-        listitem = xbmcgui.ListItem()
-        self.base = listing.Context(listitem)
+        self.listitem = xbmcgui.ListItem()
+        self.base = listing.Context()
         self.org_routes = dispatcher.registered_routes.copy()
 
         # noinspection PyUnusedLocal
@@ -350,7 +351,7 @@ class Context(unittest.TestCase):
 
     def test_close(self):
         self.base.related(self.test_callback)
-        self.base._close()
+        self.base._close(self.listitem)
 
 
 class TestListitem(unittest.TestCase):
@@ -514,3 +515,19 @@ class TestListitem(unittest.TestCase):
         self.assertTrue(listitem.art["thumb"].endswith("search.png"))
         self.assertIsInstance(listitem._args, tuple)
         self.assertTrue(len(listitem._args) == 0)
+
+    def test_pickle(self):
+        self.listitem.info["test"] = "data"
+        self.listitem.params["test"] = "data"
+        self.listitem.art["test"] = "data"
+        self.listitem.property["test"] = "data"
+        self.listitem.stream["test"] = "data"
+
+        pickled_data = pickle.dumps(self.listitem, protocol=pickle.HIGHEST_PROTOCOL)
+        new_listitem = pickle.loads(pickled_data)
+
+        self.assertEqual(new_listitem.info.get("test"), "data")
+        self.assertEqual(new_listitem.params.get("test"), "data")
+        self.assertEqual(new_listitem.art.get("test"), "data")
+        self.assertEqual(new_listitem.property.get("test"), "data")
+        self.assertEqual(new_listitem.stream.get("test"), "data")